insane speed
now you need to open the settings to view the metadata.

if you cycle through the images and open settings again, it will open the current image. (no metadata history since it's only read upon user input when opening settings)

metadata no longer scales to the right of your window, instead it wraps at a specific number that you can set and update in real time in the settings.

No more counter, which means that if you drag an insane amount of files, it won't open them all to check if they're valid. instead, the app will just not load them once the number in `filespath` reaches that file.

Also, the app no longer crashes on invalid numeric value, and resets to the default value instead.